hotfix: type 에러 수정
