--- src/domain/productDetail/productDetail.types.ts
@@ -1,12 +1,0 @@
-export interface SizeInfo {
-  size: string
-  quantity: number
-}
-
-export interface Product {
-  id: number
-  name: string
-  price: number
-  imageUrl: string
-  clothesSizeList: SizeInfo[]
-}
