+++ src/services/cartApi.ts
@@ -33,6 +33,7 @@ export const postCartItems = async (
   size: string,
   quantity: number = 1
 ): Promise<void> => {
+  console.log({ clothesId, size, quantity })
   const response: ApiResponse<void | { code: string; message: string }> =
     await axiosInstance.post('/api/v1/carts', { clothesId, size, quantity })
   if (!response.success) {
